x64 Phase 2b (3/N): scattered .cpp size_t-cast cleanup
Bulk-walked through the ~10 next files (each contributing 6-11 of
the remaining narrowing warnings) and applied the standard
(INT)/(DWORD)/(int) casts.

Files in this chunk:

  app/cedtAppConf.cpp           strlen at every magic-header read/write
  doc/cedtDocAnal.cpp           tellg() and getline buffer-fill
  doc/cedtDocEditAdv.cpp        GetCount() + strlen for line-comment LCS
  doc/cedtDocSearch.cpp         GetCount() + strlen for pair-bracket spans
  panels/FileWndDirectory.cpp   GetSize() + strlen for the local-drive
                                array and dir-tree path traversal
  panels/FileWndProject.cpp     strlen for the <item> attribute parser
                                + pointer-diff casts for CString(beg, len)
                                in attribute name/value extraction
  panels/FileTab.cpp            pMsg->wParam (WPARAM) → (UINT) cast at
                                tab-ctrl L/M-button forwarders;
                                pNMHDR->idFrom → (UINT) for tooltip id
  network/RemoteFile.cpp        CFtpFileFind::GetLength() and
                                CFileStatus::m_size narrowing to DWORD;
                                CString(BEG, FWD-BEG) span casts in
                                the ls-style file-mode/time parser;
                                also fixed a stray ';' after #include
                                "stdafx.h" that was throwing C4067
  util/encode.cpp               strlen in hex_encode / hex_decode
  util/evaluate.cpp             token-span (pEnd - pExpr) → (INT)
  util/Utility.cpp              already in chunk 2 but added a couple
                                of straggler nFound / nLocal cast
                                (covered by replace_all on the
                                "pResult - pString" pattern)
  view/cedtView.cpp             GetCount() in the four scroll handlers,
                                plus the WPARAM → UINT/TCHAR/BYTE
                                conversion at every WM_CHAR /
                                WM_KEYDOWN dispatch site (IME-aware
                                input path, so kept the existing
                                semantics — just made the narrowing
                                explicit)
  view/cedtViewCommand.cpp      arrPathName.GetSize() + CString span
                                casts in the shell-variable expander
                                + strlen in ExpandShellVariable
  view/cedtViewSearch.cpp       SEARCH_REG_EXP(...) ternary results
                                where one branch is strlen() (size_t)
                                — cast the strlen branch to (INT) so
                                the ternary type doesn't get widened

Effect on Debug-KR|x64 clean rebuild:

  before this chunk : 157 warnings
  after  this chunk :  71 warnings

Cumulative since the start of Phase 2:

  start of Phase 2 : 756 warnings
  now              :  71 warnings  (90.6% removed)

Remaining 71 break down as:

  C4244 (37) + C4267 (16) = 53  small-scattered narrowings, mostly
                                1-3 per file in ~20 files left
  C4996 (12)                    deprecated POSIX names (stricmp /
                                strnicmp / strlwr) — Phase 2c
  C5033  (2)                    'register' is no longer a storage
                                class in C++17 — Phase 2c
  C4838  (2)                    UINT → CLIPFORMAT narrowing in brace
                                init — Phase 2c (DDV-style fix)
  C4090  (1)                    const correctness — Phase 2c
  C4091  (1)                    typedef ignored — Phase 2c

cedt_kr.exe x64 still builds clean.

+++ src/app/cedtAppConf.cpp
@@ -11,7 +11,7 @@ BOOL CCedtApp::SaveFtpAccountInfo(LPCTSTR lpszPathName)
 	ofstream fout(lpszPathName, ios::out | ios::binary);
 	if( ! fout.is_open() ) return FALSE;
 
-	INT nLength = strlen(STRING_FTPACCOUNTVER);
+	INT nLength = (INT)strlen(STRING_FTPACCOUNTVER);
 	fout.write((const char *)STRING_FTPACCOUNTVER, nLength);
 
 	for(INT i = 0; i < MAX_FTP_ACCOUNT; i++) {
@@ -27,7 +27,7 @@ BOOL CCedtApp::LoadFtpAccountInfo(LPCTSTR lpszPathName)
 	ifstream fin(lpszPathName, ios::in | ios::binary);
 	if( ! fin.is_open() ) return FALSE;
 
-	TCHAR szBuffer[2048]; INT nLength = strlen(STRING_FTPACCOUNTVER); fin.read((char *)szBuffer, nLength); szBuffer[nLength] = '\0';
+	TCHAR szBuffer[2048]; INT nLength = (INT)strlen(STRING_FTPACCOUNTVER); fin.read((char *)szBuffer, nLength); szBuffer[nLength] = '\0';
 	if( strcmp(szBuffer, STRING_FTPACCOUNTVER) ) { fin.close(); return FALSE; }
 
 	for(INT i = 0; i < MAX_FTP_ACCOUNT; i++) {
@@ -45,7 +45,7 @@ BOOL CCedtApp::SaveUserCommands(LPCTSTR lpszPathName)
 	ofstream fout(lpszPathName, ios::out | ios::binary);
 	if( ! fout.is_open() ) return FALSE;
 
-	INT nLength = strlen(STRING_USERTOOLSVER);
+	INT nLength = (INT)strlen(STRING_USERTOOLSVER);
 	fout.write((const char *)STRING_USERTOOLSVER, nLength);
 
 	for(INT i = 0; i < 11; i++) {
@@ -61,7 +61,7 @@ BOOL CCedtApp::LoadUserCommands(LPCTSTR lpszPathName)
 	ifstream fin(lpszPathName, ios::in | ios::binary);
 	if( ! fin.is_open() ) return FALSE;
 
-	TCHAR szBuffer[2048]; INT nLength = strlen(STRING_USERTOOLSVER); fin.read((char *)szBuffer, nLength); szBuffer[nLength] = '\0';
+	TCHAR szBuffer[2048]; INT nLength = (INT)strlen(STRING_USERTOOLSVER); fin.read((char *)szBuffer, nLength); szBuffer[nLength] = '\0';
 	if( strcmp(szBuffer, STRING_USERTOOLSVER) ) { fin.close(); return FALSE; }
 
 	for(INT i = 0; i < 11; i++) {
@@ -79,7 +79,7 @@ BOOL CCedtApp::SaveMacroBuffers(LPCTSTR lpszPathName)
 	ofstream fout(lpszPathName, ios::out | ios::binary);
 	if( ! fout.is_open() ) return FALSE;
 
-	INT nLength = strlen(STRING_USERMACROVER);
+	INT nLength = (INT)strlen(STRING_USERMACROVER);
 	fout.write((const char *)STRING_USERMACROVER, nLength);
 
 	for(INT i = 0; i < 11; i++) {
@@ -95,7 +95,7 @@ BOOL CCedtApp::LoadMacroBuffers(LPCTSTR lpszPathName)
 	ifstream fin(lpszPathName, ios::in | ios::binary);
 	if( ! fin.is_open() ) return FALSE;
 
-	TCHAR szBuffer[2048]; INT nLength = strlen(STRING_USERMACROVER); fin.read((char *)szBuffer, nLength); szBuffer[nLength] = '\0';
+	TCHAR szBuffer[2048]; INT nLength = (INT)strlen(STRING_USERMACROVER); fin.read((char *)szBuffer, nLength); szBuffer[nLength] = '\0';
 	if( strcmp(szBuffer, STRING_USERMACROVER) ) { fin.close(); return FALSE; }
 
 	for(INT i = 0; i < 11; i++) {
@@ -294,7 +294,7 @@ BOOL CCedtApp::LoadColorScheme(LPCTSTR lpszPathName)
 	ifstream fin(lpszPathName, ios::in | ios::binary);
 	if( ! fin.is_open() ) return FALSE;
 
-	TCHAR szBuffer[2048]; INT nLength = strlen(STRING_COLORSETTINGSVER); fin.read((char *)szBuffer, nLength); szBuffer[nLength] = '\0';
+	TCHAR szBuffer[2048]; INT nLength = (INT)strlen(STRING_COLORSETTINGSVER); fin.read((char *)szBuffer, nLength); szBuffer[nLength] = '\0';
 	if( strcmp(szBuffer, STRING_COLORSETTINGSVER) ) { fin.close(); return FALSE; }
 
 	fin.read((char *)CCedtView::m_crBkgrColor, sizeof(CCedtView::m_crBkgrColor));
@@ -315,7 +315,7 @@ BOOL CCedtApp::SaveColorScheme(LPCTSTR lpszPathName)
 	ofstream fout(lpszPathName, ios::out | ios::binary);
 	if( ! fout.is_open() ) return FALSE;
 
-	INT nLength = strlen(STRING_COLORSETTINGSVER);
+	INT nLength = (INT)strlen(STRING_COLORSETTINGSVER);
 	fout.write((const char *)STRING_COLORSETTINGSVER, nLength);
 
 	fout.write((const char *)CCedtView::m_crBkgrColor, sizeof(CCedtView::m_crBkgrColor));
@@ -503,7 +503,7 @@ BOOL CCedtApp::LoadUserConfiguration(LPCTSTR lpszPathName)
 	ifstream fin(lpszPathName, ios::in | ios::binary);
 	if( ! fin.is_open() ) return FALSE;
 
-	nLength = strlen(STRING_CONFIGURATIONVER); fin.read((char *)szBuffer, nLength); szBuffer[nLength] = '\0';
+	nLength = (INT)strlen(STRING_CONFIGURATIONVER); fin.read((char *)szBuffer, nLength); szBuffer[nLength] = '\0';
 	if( strcmp(szBuffer, STRING_CONFIGURATIONVER) ) { fin.close(); return FALSE; }
 
 	fin.read((char *)CCedtView::m_lfScreen, sizeof(CCedtView::m_lfScreen));
@@ -705,7 +705,7 @@ BOOL CCedtApp::SaveUserConfiguration(LPCTSTR lpszPathName)
 	ofstream fout(lpszPathName, ios::out | ios::binary);
 	if( ! fout.is_open() ) return FALSE;
 
-	nLength = strlen(STRING_CONFIGURATIONVER);
+	nLength = (INT)strlen(STRING_CONFIGURATIONVER);
 	fout.write((const char *)STRING_CONFIGURATIONVER, nLength);
 
 	fout.write((const char *)CCedtView::m_lfScreen, sizeof(CCedtView::m_lfScreen));
